requirements.txt
langchain_openai
langchain_core
python_dotenv
streamlit
langchain_community
langserve